Moved blub bundle validation to common
