+++ Cargo.toml
@@ -65,3 +65,4 @@ cfg-if = "1.0.0"
 reqwest = { version = "0.12.7", features = ["json"] }
 snap = "1.1.1"
 c-kzg = "^1.0.3"
+k256 = { version = "0.13.3", features = ["ecdh"] }
