+++ cmd/ethereum_rust/Cargo.toml
@@ -19,13 +19,13 @@ bytes.workspace = true
 hex.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
+k256.workspace = true
 clap = { version = "4.5.4", features = ["cargo"] }
 directories = "5.0.1"
 serde_json.workspace = true
 tokio = { version = "1.38.0", features = ["full"] }
 anyhow = "1.0.86"
 rand = "0.8.5"
-k256 = { version = "0.13.3", features = ["ecdh"] }
 local-ip-address = "0.6"
 tokio-util.workspace = true
