+++ crates/blockchain/Cargo.toml
@@ -10,7 +10,6 @@ thiserror.workspace = true
 sha3.workspace = true
 tracing.workspace = true
 bytes.workspace = true
-c-kzg.workspace = true
 cfg-if = "1.0.0"
 
 ethereum_rust-rlp.workspace = true
