+++ crates/blockchain/mempool.rs
@@ -5,7 +5,7 @@ use crate::{
         MAX_INITCODE_SIZE, MIN_BASE_FEE_PER_BLOB_GAS, TX_ACCESS_LIST_ADDRESS_GAS,
         TX_ACCESS_LIST_STORAGE_KEY_GAS, TX_CREATE_GAS_COST, TX_DATA_NON_ZERO_GAS,
         TX_DATA_NON_ZERO_GAS_EIP2028, TX_DATA_ZERO_GAS_COST, TX_GAS_COST,
-        TX_INIT_CODE_WORD_GAS_COST, VERSIONED_HASH_VERSION_KZG,
+        TX_INIT_CODE_WORD_GAS_COST,
     },
     error::MempoolError,
 };
@@ -27,7 +27,7 @@ pub fn add_blob_transaction(
     validate_blob_transaction(&transaction, &blobs_bundle)?;
 
     // validate blobs bundle
-    validate_blobs_bundle(&transaction, &blobs_bundle)?;
+    blobs_bundle.validate(&transaction)?;
 
     // Validate transaction
     let transaction = Transaction::EIP4844Transaction(transaction);
@@ -258,31 +258,6 @@ fn validate_blob_transaction(
     Ok(())
 }
 
-fn validate_blobs_bundle(
-    tx: &EIP4844Transaction,
-    blobs_bundle: &BlobsBundle,
-) -> Result<(), MempoolError> {
-    // return error early if any commitment doesn't match it's blob versioned hash
-    for (commitment, blob_versioned_hash) in blobs_bundle
-        .commitments
-        .iter()
-        .zip(tx.blob_versioned_hashes.iter())
-    {
-        if *blob_versioned_hash != kzg_to_versioned_hash(commitment) {
-            return Err(MempoolError::BlobVersionedHashesIncorrectError);
-        }
-    }
-
-    Ok(())
-}
-
-fn kzg_to_versioned_hash(data: &[u8]) -> H256 {
-    use k256::sha2::Digest;
-    let mut versioned_hash: [u8; 32] = k256::sha2::Sha256::digest(data).into();
-    versioned_hash[0] = VERSIONED_HASH_VERSION_KZG;
-    versioned_hash.into()
-}
-
 fn transaction_intrinsic_gas(
     tx: &Transaction,
     header: &BlockHeader,
@@ -352,7 +327,7 @@ mod tests {
         TX_DATA_ZERO_GAS_COST, TX_GAS_COST, TX_INIT_CODE_WORD_GAS_COST,
     };
 
-    use super::{transaction_intrinsic_gas, validate_blobs_bundle, validate_transaction};
+    use super::{transaction_intrinsic_gas, validate_transaction};
     use ethereum_rust_core::types::{
         BlobsBundle, BlockHeader, ChainConfig, EIP1559Transaction, EIP4844Transaction, Transaction,
         TxKind, BYTES_PER_BLOB,
@@ -700,118 +675,6 @@ mod tests {
         assert!(matches!(
             validation,
             Err(MempoolError::TxBlobBaseFeeTooLowError)
-        ));
-    }
-
-    #[test]
-    fn transaction_with_correct_blobs_should_pass() {
-        let convert_str_to_bytes48 = |s| {
-            let bytes = hex::decode(s).expect("Invalid hex string");
-            let mut array = [0u8; 48];
-            array.copy_from_slice(&bytes[..48]);
-            array
-        };
-
-        // blob data taken from: https://etherscan.io/tx/0x02a623925c05c540a7633ffa4eb78474df826497faa81035c4168695656801a2#blobs
-
-        let blobs_bundle = BlobsBundle {
-            blobs: vec![[0; BYTES_PER_BLOB], [0; BYTES_PER_BLOB]],
-            commitments: vec!["b90289aabe0fcfb8db20a76b863ba90912d1d4d040cb7a156427d1c8cd5825b4d95eaeb221124782cc216960a3d01ec5",
-                              "91189a03ce1fe1225fc5de41d502c3911c2b19596f9011ea5fca4bf311424e5f853c9c46fe026038036c766197af96a0"]
-                              .into_iter()
-                              .map(|s| {
-                                  convert_str_to_bytes48(s)
-                              })
-                              .collect(),
-            proofs: vec!["b502263fc5e75b3587f4fb418e61c5d0f0c18980b4e00179326a65d082539a50c063507a0b028e2db10c55814acbe4e9",
-                         "a29c43f6d05b7f15ab6f3e5004bd5f6b190165dc17e3d51fd06179b1e42c7aef50c145750d7c1cd1cd28357593bc7658"]
-                         .into_iter()
-                              .map(|s| {
-                                  convert_str_to_bytes48(s)
-                              })
-                              .collect()
-        };
-
-        let tx = EIP4844Transaction {
-            nonce: 3,
-            max_priority_fee_per_gas: 0,
-            max_fee_per_gas: 0,
-            max_fee_per_blob_gas: 0.into(),
-            gas: 15_000_000,
-            to: Address::from_low_u64_be(1), // Normal tx
-            value: U256::zero(),             // Value zero
-            data: Bytes::default(),          // No data
-            access_list: Default::default(), // No access list
-            blob_versioned_hashes: vec![
-                "01ec8054d05bfec80f49231c6e90528bbb826ccd1464c255f38004099c8918d9",
-                "0180cb2dee9e6e016fabb5da4fb208555f5145c32895ccd13b26266d558cd77d",
-            ]
-            .into_iter()
-            .map(|b| {
-                let bytes = hex::decode(b).expect("Invalid hex string");
-                H256::from_slice(&bytes)
-            })
-            .collect::<Vec<H256>>(),
-            ..Default::default()
-        };
-
-        assert!(matches!(validate_blobs_bundle(&tx, &blobs_bundle), Ok(())));
-    }
-
-    #[test]
-    fn transaction_with_incorrect_blobs_should_fail() {
-        let convert_str_to_bytes48 = |s| {
-            let bytes = hex::decode(s).expect("Invalid hex string");
-            let mut array = [0u8; 48];
-            array.copy_from_slice(&bytes[..48]);
-            array
-        };
-
-        // blob data taken from: https://etherscan.io/tx/0x02a623925c05c540a7633ffa4eb78474df826497faa81035c4168695656801a2#blobs
-        let blobs_bundle = BlobsBundle {
-            blobs: vec![[0; BYTES_PER_BLOB], [0; BYTES_PER_BLOB]],
-            commitments: vec!["dead89aabe0fcfb8db20a76b863ba90912d1d4d040cb7a156427d1c8cd5825b4d95eaeb221124782cc216960a3d01ec5",
-                              "91189a03ce1fe1225fc5de41d502c3911c2b19596f9011ea5fca4bf311424e5f853c9c46fe026038036c766197af96a0"]
-                              .into_iter()
-                              .map(|s| {
-                                  convert_str_to_bytes48(s)
-                              })
-                              .collect(),
-            proofs: vec!["b502263fc5e75b3587f4fb418e61c5d0f0c18980b4e00179326a65d082539a50c063507a0b028e2db10c55814acbe4e9",
-                         "a29c43f6d05b7f15ab6f3e5004bd5f6b190165dc17e3d51fd06179b1e42c7aef50c145750d7c1cd1cd28357593bc7658"]
-                         .into_iter()
-                              .map(|s| {
-                                  convert_str_to_bytes48(s)
-                              })
-                              .collect()
-        };
-
-        let tx = EIP4844Transaction {
-            nonce: 3,
-            max_priority_fee_per_gas: 0,
-            max_fee_per_gas: 0,
-            max_fee_per_blob_gas: 0.into(),
-            gas: 15_000_000,
-            to: Address::from_low_u64_be(1), // Normal tx
-            value: U256::zero(),             // Value zero
-            data: Bytes::default(),          // No data
-            access_list: Default::default(), // No access list
-            blob_versioned_hashes: vec![
-                "01ec8054d05bfec80f49231c6e90528bbb826ccd1464c255f38004099c8918d9",
-                "0180cb2dee9e6e016fabb5da4fb208555f5145c32895ccd13b26266d558cd77d",
-            ]
-            .into_iter()
-            .map(|b| {
-                let bytes = hex::decode(b).expect("Invalid hex string");
-                H256::from_slice(&bytes)
-            })
-            .collect::<Vec<H256>>(),
-            ..Default::default()
-        };
-
-        assert!(matches!(
-            validate_blobs_bundle(&tx, &blobs_bundle),
-            Err(MempoolError::BlobVersionedHashesIncorrectError)
         ));
     }
 }
