+++ crates/common/Cargo.toml
@@ -14,6 +14,9 @@ ethereum-types.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
+k256.workspace = true
+c-kzg.workspace = true
+
 keccak-hash = "0.10.0"
 sha3.workspace = true
 secp256k1 = { version = "0.29", default-features = false, features = [
