+++ crates/common/types/blobs_bundle.rs
@@ -1,6 +1,13 @@
 use std::ops::AddAssign;
 
 use crate::serde_utils;
+use crate::{
+    types::{
+        transaction::EIP4844Transaction,
+        constants::VERSIONED_HASH_VERSION_KZG,
+    },
+    H256,
+};
 use ethereum_rust_rlp::{
     decode::RLPDecode,
     encode::RLPEncode,
@@ -26,6 +33,30 @@ pub struct BlobsBundle {
     pub commitments: Vec<Commitment>,
     #[serde(with = "serde_utils::bytes48::vec")]
     pub proofs: Vec<Proof>,
+}
+
+fn kzg_commitment_to_versioned_hash(data: &Commitment) -> H256 {
+    use k256::sha2::Digest;
+    let mut versioned_hash: [u8; 32] = k256::sha2::Sha256::digest(data).into();
+    versioned_hash[0] = VERSIONED_HASH_VERSION_KZG;
+    versioned_hash.into()
+}
+
+impl BlobsBundle {
+    pub fn validate(&self, tx: &EIP4844Transaction) -> Result<(), BlobsBundleError> {
+        // return error early if any commitment doesn't match it's blob versioned hash
+        for (commitment, blob_versioned_hash) in self
+            .commitments
+            .iter()
+            .zip(tx.blob_versioned_hashes.iter())
+        {
+            if *blob_versioned_hash != kzg_commitment_to_versioned_hash(&commitment) {
+                return Err(BlobsBundleError::BlobIncorrectVersionedHashes);
+            }
+        }
+
+        Ok(())
+    }
 }
 
 impl RLPEncode for BlobsBundle {
@@ -63,3 +94,133 @@ impl AddAssign for BlobsBundle {
         self.proofs.extend_from_slice(&rhs.proofs);
     }
 }
+
+#[derive(Debug, thiserror::Error)]
+pub enum BlobsBundleError {
+    #[error("Blob versioned hashes are incorrect")]
+    BlobIncorrectVersionedHashes,
+}
+
+#[cfg(test)]
+
+mod tests {
+    use super::*;
+    use crate::{
+        types::transaction::EIP4844Transaction,
+        Address, U256, Bytes,
+    };
+
+    #[test]
+    fn transaction_with_correct_blobs_should_pass() {
+        let convert_str_to_bytes48 = |s| {
+            let bytes = hex::decode(s).expect("Invalid hex string");
+            let mut array = [0u8; 48];
+            array.copy_from_slice(&bytes[..48]);
+            array
+        };
+
+        // blob data taken from: https://etherscan.io/tx/0x02a623925c05c540a7633ffa4eb78474df826497faa81035c4168695656801a2#blobs
+
+        let blobs_bundle = BlobsBundle {
+            blobs: vec![[0; BYTES_PER_BLOB], [0; BYTES_PER_BLOB]],
+            commitments: vec!["b90289aabe0fcfb8db20a76b863ba90912d1d4d040cb7a156427d1c8cd5825b4d95eaeb221124782cc216960a3d01ec5",
+                              "91189a03ce1fe1225fc5de41d502c3911c2b19596f9011ea5fca4bf311424e5f853c9c46fe026038036c766197af96a0"]
+                              .into_iter()
+                              .map(|s| {
+                                  convert_str_to_bytes48(s)
+                              })
+                              .collect(),
+            proofs: vec!["b502263fc5e75b3587f4fb418e61c5d0f0c18980b4e00179326a65d082539a50c063507a0b028e2db10c55814acbe4e9",
+                         "a29c43f6d05b7f15ab6f3e5004bd5f6b190165dc17e3d51fd06179b1e42c7aef50c145750d7c1cd1cd28357593bc7658"]
+                         .into_iter()
+                              .map(|s| {
+                                  convert_str_to_bytes48(s)
+                              })
+                              .collect()
+        };
+
+        let tx = EIP4844Transaction {
+            nonce: 3,
+            max_priority_fee_per_gas: 0,
+            max_fee_per_gas: 0,
+            max_fee_per_blob_gas: 0.into(),
+            gas: 15_000_000,
+            to: Address::from_low_u64_be(1), // Normal tx
+            value: U256::zero(),             // Value zero
+            data: Bytes::default(),          // No data
+            access_list: Default::default(), // No access list
+            blob_versioned_hashes: vec![
+                "01ec8054d05bfec80f49231c6e90528bbb826ccd1464c255f38004099c8918d9",
+                "0180cb2dee9e6e016fabb5da4fb208555f5145c32895ccd13b26266d558cd77d",
+            ]
+            .into_iter()
+            .map(|b| {
+                let bytes = hex::decode(b).expect("Invalid hex string");
+                H256::from_slice(&bytes)
+            })
+            .collect::<Vec<H256>>(),
+            ..Default::default()
+        };
+
+        assert!(matches!(blobs_bundle.validate(&tx), Ok(())));
+    }
+
+    #[test]
+    fn transaction_with_incorrect_blobs_should_fail() {
+        let convert_str_to_bytes48 = |s| {
+            let bytes = hex::decode(s).expect("Invalid hex string");
+            let mut array = [0u8; 48];
+            array.copy_from_slice(&bytes[..48]);
+            array
+        };
+
+        // blob data taken from: https://etherscan.io/tx/0x02a623925c05c540a7633ffa4eb78474df826497faa81035c4168695656801a2#blobs
+        let blobs_bundle = BlobsBundle {
+            blobs: vec![[0; BYTES_PER_BLOB], [0; BYTES_PER_BLOB]],
+            commitments: vec!["dead89aabe0fcfb8db20a76b863ba90912d1d4d040cb7a156427d1c8cd5825b4d95eaeb221124782cc216960a3d01ec5",
+                              "91189a03ce1fe1225fc5de41d502c3911c2b19596f9011ea5fca4bf311424e5f853c9c46fe026038036c766197af96a0"]
+                              .into_iter()
+                              .map(|s| {
+                                  convert_str_to_bytes48(s)
+                              })
+                              .collect(),
+            proofs: vec!["b502263fc5e75b3587f4fb418e61c5d0f0c18980b4e00179326a65d082539a50c063507a0b028e2db10c55814acbe4e9",
+                         "a29c43f6d05b7f15ab6f3e5004bd5f6b190165dc17e3d51fd06179b1e42c7aef50c145750d7c1cd1cd28357593bc7658"]
+                         .into_iter()
+                              .map(|s| {
+                                  convert_str_to_bytes48(s)
+                              })
+                              .collect()
+        };
+
+        let tx = EIP4844Transaction {
+            nonce: 3,
+            max_priority_fee_per_gas: 0,
+            max_fee_per_gas: 0,
+            max_fee_per_blob_gas: 0.into(),
+            gas: 15_000_000,
+            to: Address::from_low_u64_be(1), // Normal tx
+            value: U256::zero(),             // Value zero
+            data: Bytes::default(),          // No data
+            access_list: Default::default(), // No access list
+            blob_versioned_hashes: vec![
+                "01ec8054d05bfec80f49231c6e90528bbb826ccd1464c255f38004099c8918d9",
+                "0180cb2dee9e6e016fabb5da4fb208555f5145c32895ccd13b26266d558cd77d",
+            ]
+            .into_iter()
+            .map(|b| {
+                let bytes = hex::decode(b).expect("Invalid hex string");
+                H256::from_slice(&bytes)
+            })
+            .collect::<Vec<H256>>(),
+            ..Default::default()
+        };
+
+        assert!(matches!(
+            blobs_bundle.validate(&tx),
+            Err(BlobsBundleError::BlobIncorrectVersionedHashes)
+        ));
+    }
+}
+ 
+
